cleanup a lot of mixins

+++ src/main/java/luke/bonusblocks/mixin/ItemMixin.java
@@ -47,7 +47,7 @@ public class ItemMixin {
 
 					if (newMeta < 3) {
 						world.setBlockAndMetadataWithNotify(blockX, blockY, blockZ, BonusBlocks.OVERLAY_RAW_IRON.id(), newMeta);
-						world.playBlockSoundEffect(player, (float) blockX + 0.5F, (float) blockY + 0.5F, (float) blockZ + 0.5F, BonusBlocks.OVERLAY_RAW_IRON, EnumBlockSoundEffectType.PLACE);
+						world.playBlockSoundEffect(player, blockX + 0.5F, blockY + 0.5F, blockZ + 0.5F, BonusBlocks.OVERLAY_RAW_IRON, EnumBlockSoundEffectType.PLACE);
 						itemstack.consumeItem(player);
 						player.swingItem();
 						cir.setReturnValue(true);
@@ -64,14 +64,14 @@ public class ItemMixin {
 
 				if (id == BonusBlocks.OVERLAY_RAW_IRON.id()) {
 					newMeta = meta + 1;
-					AABB bbBox = AABB.getTemporaryBB(blockX, blockY, blockZ, (float) blockX + 1.0F, (float) blockY + (float) (2 * (newMeta + 1)) / 16.0F, (float) blockZ + 1.0F);
+					AABB bbBox = AABB.getTemporaryBB(blockX, blockY, blockZ, blockX + 1.0F, blockY + (2 * (newMeta + 1)) / 16.0F, blockZ + 1.0F);
 					if (!world.checkIfAABBIsClear(bbBox) || !world.isBlockOpaqueCube(blockX, blockY - 1, blockZ)) {
 						cir.setReturnValue(false);
 					}
 
 					if (newMeta < 3) {
 						world.setBlockAndMetadataWithNotify(blockX, blockY, blockZ, BonusBlocks.OVERLAY_RAW_IRON.id(), newMeta);
-						world.playBlockSoundEffect(player, (float) blockX + 0.5F, (float) blockY + 0.5F, (float) blockZ + 0.5F, BonusBlocks.OVERLAY_RAW_IRON, EnumBlockSoundEffectType.PLACE);
+						world.playBlockSoundEffect(player, blockX + 0.5F, blockY + 0.5F, blockZ + 0.5F, BonusBlocks.OVERLAY_RAW_IRON, EnumBlockSoundEffectType.PLACE);
 						itemstack.consumeItem(player);
 						player.swingItem();
 						cir.setReturnValue(true);
@@ -80,7 +80,7 @@ public class ItemMixin {
 
 				if (world.canBlockBePlacedAt(BonusBlocks.OVERLAY_RAW_IRON.id(), blockX, blockY, blockZ, false, side) && world.isBlockOpaqueCube(blockX, blockY - 1, blockZ) && world.setBlockAndMetadataWithNotify(blockX, blockY, blockZ, BonusBlocks.OVERLAY_RAW_IRON.id(), 0)) {
 					BonusBlocks.OVERLAY_RAW_IRON.onBlockPlacedByMob(world, blockX, blockY, blockZ, side, player, xPlaced, yPlaced);
-					world.playBlockSoundEffect(player, (float) blockX + 0.5F, (float) blockY + 0.5F, (float) blockZ + 0.5F, BonusBlocks.OVERLAY_RAW_IRON, EnumBlockSoundEffectType.PLACE);
+					world.playBlockSoundEffect(player, blockX + 0.5F, blockY + 0.5F, blockZ + 0.5F, BonusBlocks.OVERLAY_RAW_IRON, EnumBlockSoundEffectType.PLACE);
 					itemstack.consumeItem(player);
 					player.swingItem();
 					cir.setReturnValue(true);
@@ -112,7 +112,7 @@ public class ItemMixin {
 
 					if (newMeta < 3) {
 						world.setBlockAndMetadataWithNotify(blockX, blockY, blockZ, BonusBlocks.OVERLAY_RAW_GOLD.id(), newMeta);
-						world.playBlockSoundEffect(player, (float) blockX + 0.5F, (float) blockY + 0.5F, (float) blockZ + 0.5F, BonusBlocks.OVERLAY_RAW_GOLD, EnumBlockSoundEffectType.PLACE);
+						world.playBlockSoundEffect(player, blockX + 0.5F, blockY + 0.5F, blockZ + 0.5F, BonusBlocks.OVERLAY_RAW_GOLD, EnumBlockSoundEffectType.PLACE);
 						itemstack.consumeItem(player);
 						player.swingItem();
 						cir.setReturnValue(true);
@@ -129,14 +129,14 @@ public class ItemMixin {
 
 				if (id == BonusBlocks.OVERLAY_RAW_GOLD.id()) {
 					newMeta = meta + 1;
-					AABB bbBox = AABB.getTemporaryBB(blockX, blockY, blockZ, (float) blockX + 1.0F, (float) blockY + (float) (2 * (newMeta + 1)) / 16.0F, (float) blockZ + 1.0F);
+					AABB bbBox = AABB.getTemporaryBB(blockX, blockY, blockZ, blockX + 1.0F, blockY + (2 * (newMeta + 1)) / 16.0F, blockZ + 1.0F);
 					if (!world.checkIfAABBIsClear(bbBox) || !world.isBlockOpaqueCube(blockX, blockY - 1, blockZ)) {
 						cir.setReturnValue(false);
 					}
 
 					if (newMeta < 3) {
 						world.setBlockAndMetadataWithNotify(blockX, blockY, blockZ, BonusBlocks.OVERLAY_RAW_GOLD.id(), newMeta);
-						world.playBlockSoundEffect(player, (float) blockX + 0.5F, (float) blockY + 0.5F, (float) blockZ + 0.5F, BonusBlocks.OVERLAY_RAW_GOLD, EnumBlockSoundEffectType.PLACE);
+						world.playBlockSoundEffect(player, blockX + 0.5F, blockY + 0.5F, blockZ + 0.5F, BonusBlocks.OVERLAY_RAW_GOLD, EnumBlockSoundEffectType.PLACE);
 						itemstack.consumeItem(player);
 						player.swingItem();
 						cir.setReturnValue(true);
@@ -145,7 +145,7 @@ public class ItemMixin {
 
 				if (world.canBlockBePlacedAt(BonusBlocks.OVERLAY_RAW_GOLD.id(), blockX, blockY, blockZ, false, side) && world.isBlockOpaqueCube(blockX, blockY - 1, blockZ) && world.setBlockAndMetadataWithNotify(blockX, blockY, blockZ, BonusBlocks.OVERLAY_RAW_GOLD.id(), 0)) {
 					BonusBlocks.OVERLAY_RAW_GOLD.onBlockPlacedByMob(world, blockX, blockY, blockZ, side, player, xPlaced, yPlaced);
-					world.playBlockSoundEffect(player, (float) blockX + 0.5F, (float) blockY + 0.5F, (float) blockZ + 0.5F, BonusBlocks.OVERLAY_RAW_GOLD, EnumBlockSoundEffectType.PLACE);
+					world.playBlockSoundEffect(player, blockX + 0.5F, blockY + 0.5F, blockZ + 0.5F, BonusBlocks.OVERLAY_RAW_GOLD, EnumBlockSoundEffectType.PLACE);
 					itemstack.consumeItem(player);
 					player.swingItem();
 					cir.setReturnValue(true);
@@ -177,7 +177,7 @@ public class ItemMixin {
 
 					if (newMeta < 3) {
 						world.setBlockAndMetadataWithNotify(blockX, blockY, blockZ, BonusBlocks.OVERLAY_FLINT.id(), newMeta);
-						world.playBlockSoundEffect(player, (float) blockX + 0.5F, (float) blockY + 0.5F, (float) blockZ + 0.5F, BonusBlocks.OVERLAY_FLINT, EnumBlockSoundEffectType.PLACE);
+						world.playBlockSoundEffect(player, blockX + 0.5F, blockY + 0.5F, blockZ + 0.5F, BonusBlocks.OVERLAY_FLINT, EnumBlockSoundEffectType.PLACE);
 						itemstack.consumeItem(player);
 						player.swingItem();
 						cir.setReturnValue(true);
@@ -194,14 +194,14 @@ public class ItemMixin {
 
 				if (id == BonusBlocks.OVERLAY_FLINT.id()) {
 					newMeta = meta + 1;
-					AABB bbBox = AABB.getTemporaryBB(blockX, blockY, blockZ, (float) blockX + 1.0F, (float) blockY + (float) (2 * (newMeta + 1)) / 16.0F, (float) blockZ + 1.0F);
+					AABB bbBox = AABB.getTemporaryBB(blockX, blockY, blockZ, blockX + 1.0F, blockY + (2 * (newMeta + 1)) / 16.0F, blockZ + 1.0F);
 					if (!world.checkIfAABBIsClear(bbBox) || !world.isBlockOpaqueCube(blockX, blockY - 1, blockZ)) {
 						cir.setReturnValue(false);
 					}
 
 					if (newMeta < 3) {
 						world.setBlockAndMetadataWithNotify(blockX, blockY, blockZ, BonusBlocks.OVERLAY_FLINT.id(), newMeta);
-						world.playBlockSoundEffect(player, (float) blockX + 0.5F, (float) blockY + 0.5F, (float) blockZ + 0.5F, BonusBlocks.OVERLAY_FLINT, EnumBlockSoundEffectType.PLACE);
+						world.playBlockSoundEffect(player, blockX + 0.5F, blockY + 0.5F, blockZ + 0.5F, BonusBlocks.OVERLAY_FLINT, EnumBlockSoundEffectType.PLACE);
 						itemstack.consumeItem(player);
 						player.swingItem();
 						cir.setReturnValue(true);
@@ -210,7 +210,7 @@ public class ItemMixin {
 
 				if (world.canBlockBePlacedAt(BonusBlocks.OVERLAY_FLINT.id(), blockX, blockY, blockZ, false, side) && world.isBlockOpaqueCube(blockX, blockY - 1, blockZ) && world.setBlockAndMetadataWithNotify(blockX, blockY, blockZ, BonusBlocks.OVERLAY_FLINT.id(), 0)) {
 					BonusBlocks.OVERLAY_FLINT.onBlockPlacedByMob(world, blockX, blockY, blockZ, side, player, xPlaced, yPlaced);
-					world.playBlockSoundEffect(player, (float) blockX + 0.5F, (float) blockY + 0.5F, (float) blockZ + 0.5F, BonusBlocks.OVERLAY_FLINT, EnumBlockSoundEffectType.PLACE);
+					world.playBlockSoundEffect(player, blockX + 0.5F, blockY + 0.5F, blockZ + 0.5F, BonusBlocks.OVERLAY_FLINT, EnumBlockSoundEffectType.PLACE);
 					itemstack.consumeItem(player);
 					player.swingItem();
 					cir.setReturnValue(true);
